Time: 8 ms (98.66%), Space: 25.8 MB (6.32%) - LeetHub

+++ 0530-minimum-absolute-difference-in-bst/0530-minimum-absolute-difference-in-bst.cpp
@@ -11,7 +11,7 @@
  */
 class Solution {
 public:
-     vector<int>v;
+     vector<long long int>v;
     void inorder(TreeNode *root)
     {
         if(!root)
@@ -24,7 +24,7 @@ public:
     int getMinimumDifference(TreeNode* root) {
         inorder(root);
        
-        int mn;
+        long long int mn;
         for(int i=1;i<v.size();i++)
         {
             mn=min((v[i]-v[i-1]),mn);
